психоботик/psycho_bot/requirements.txt
python-telegram-bot==22.3
python-dotenv==1.1.1
pillow==11.3.0
requests==2.32.5